WeatherApp/FypApp/Spark/requirements.txt
numpy
pandas
requests
Pillow
plotly
streamlit
pycountry
confluent-kafka==1.7.0
python-dateutil==2.8.2